Only focus on task title on initial create
