+++ lotti/lib/widgets/journal/entry_detail_widget.dart
@@ -21,7 +21,7 @@ import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/journal/linked_duration.dart';
 import 'package:lotti/widgets/journal/tags_widget.dart';
 import 'package:lotti/widgets/misc/survey_summary.dart';
-import 'package:lotti/widgets/pages/add/new_task_page.dart';
+import 'package:lotti/widgets/tasks/task_form.dart';
 import 'package:path_provider/path_provider.dart';
 
 class EntryDetailWidget extends StatefulWidget {
