+++ lotti/pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.4.23+436
+version: 0.4.24+437
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
